+++ iOSReferenceApp/Resources/Layouts/HorizontalScrollRow.swift
@@ -57,10 +57,15 @@ extension HorizontalScrollRow: UICollectionViewDelegate {
         if let preview = cell as? AssetPreviewCell {
             let vm = viewModel.content[indexPath.row]
             
+            let resizeProcessor = ResizingImageProcessor(referenceSize: viewModel.preferredThumbnailSize, mode: ContentMode.aspectFill)
+            let croppingProcessor = CroppingImageProcessor(size: viewModel.preferredThumbnailSize)
+            let roundedRectProcessor = RoundCornerImageProcessor(cornerRadius: 6)
+            let processor = (resizeProcessor>>croppingProcessor)>>roundedRectProcessor
             preview.reset()
             preview.thumbnail(title: vm.anyTitle(locale: "en"))
             if let url = imageUrl(for: indexPath) {
-                preview.thumbnailView.kf.setImage(with: url) { (image, error, _, url) in
+                preview.thumbnailView.kf.setImage(with: url, options: [.processor(processor)]) { (image, error, _, url) in
+//                preview.thumbnailView.kf.setImage(with: url) { (image, error, _, url) in
                     if let error = error {
                         print("Kingfisher: ",error)
                     }
@@ -73,7 +78,7 @@ extension HorizontalScrollRow: UICollectionViewDelegate {
         return viewModel
             .content[indexPath.row]
             .images(locale: "en")
-            .prefere(orientation: .landscape)
+            .prefere(orientation: .portrait)
             .validImageUrls()
             .first
     }