+++ iOSReferenceApp/Views/Home/AssetDetailsViewController.swift
@@ -59,8 +59,12 @@ extension AssetDetailsPresenter {
 }
 
 class AssetDetailsViewController: UIViewController {
-
+    
+    @IBOutlet weak var scrollView: LazyScrollView!
+    @IBOutlet weak var contentStackView: UIStackView!
+    
     @IBOutlet weak var mainImageView: UIImageView!
+    @IBOutlet weak var gradientView: GradientView!
     @IBOutlet weak var titleLabel: UILabel!
     
     @IBOutlet weak var ratingStarStackView: UIStackView!
@@ -73,6 +77,7 @@ class AssetDetailsViewController: UIViewController {
     @IBOutlet weak var progressBar: UIProgressView!
     @IBOutlet weak var durationLabel: UILabel!
     
+    @IBOutlet weak var participantsStackView: UIStackView!
     
     @IBOutlet weak var descriptionTextLabel: UILabel!
     @IBOutlet weak var footerTextLabel: UILabel!
@@ -103,7 +108,7 @@ class AssetDetailsViewController: UIViewController {
         super.viewDidLoad()
 
         defer { refreshUserDataUI() }
-        
+        loadAssetMetaData()
         determineDownloadUIForAsset()
     }
     
@@ -166,6 +171,11 @@ class AssetDetailsViewController: UIViewController {
         
         self.performSegue(withIdentifier: Segue.segueDetailsToPlayer.rawValue, sender: assetId)
     }
+    
+    override func viewDidLayoutSubviews() {
+        super.viewDidLayoutSubviews()
+        scrollView.contentSize = CGSize(width: contentStackView.frame.width, height: contentStackView.frame.height)
+    }
 }
 
 extension AssetDetailsViewController {
@@ -201,20 +211,19 @@ extension AssetDetailsViewController {
 
 // MARK: - User Data
 extension AssetDetailsViewController {
-    func refreshUserDataUI() {
+    func loadAssetMetaData() {
         let locale = "en"
         if let imageUrl = viewModel
             .images(locale: locale)
-            .prefere(orientation: .landscape)
+            .prefere(orientation: .square)
             .validImageUrls()
             .first {
-            mainImageView.kf.setImage(with: imageUrl) { (_, error, _, _) in
+            mainImageView.kf.setImage(with: imageUrl) { [weak self] (image, error, _, _) in
                 if let error = error {
                     print("Kingfisher error: ",error)
                 }
             }
         }
-        
         
         titleLabel.text = viewModel.anyTitle(locale: locale)
         descriptionTextLabel.text = viewModel.anyDescription(locale: locale)
@@ -224,6 +233,9 @@ extension AssetDetailsViewController {
         
         parentalRatingLabel.text = viewModel.anyParentalRating(locale: locale)
         
+        assignParticipants()
+    }
+    func refreshUserDataUI() {
         // Update last viewed progress
         update(lastViewedOffset: viewModel.lastViewedOffset)
     }
@@ -234,6 +246,38 @@ extension AssetDetailsViewController {
         progressBar.setProgress(lastViewedOffset?.progress ?? 0, animated: false)
         durationLabel.text = lastViewedOffset?.duration
     }
+    
+    private func assignParticipants() {
+        let current = participantsStackView.subviews
+        current.forEach{
+            participantsStackView.removeArrangedSubview($0)
+            $0.removeFromSuperview()
+        }
+        
+        viewModel.participantGroups()
+            .sorted{ $0.function < $1.function }
+            .map{ group -> UILabel in
+                let label = UILabel()
+                let function = (group.function + (group.names.count > 1 ? "s: " : ": ")).capitalized
+                
+                let names = group.names.joined(separator: ", ")
+                
+                label.font = UIFont(name: "Helvetica Neue", size: 14)
+                label.textColor = Color.lightGray
+                label.lineBreakMode = NSLineBreakMode.byWordWrapping
+                label.numberOfLines = 0
+                label.text = function + names
+                return label
+            }
+            .forEach{
+                participantsStackView.addArrangedSubview($0)
+        }
+        participantsStackView.layoutIfNeeded()
+    }
+}
+
+extension UIImageView {
+    
 }
 
 // MARK: - Download available